+++ yiilian-dht/src/data/find_node.rs
@@ -19,7 +19,7 @@ pub struct FindNode {
     pub ip: Option<SocketAddr>,
 
     /// readonly
-    pub ro: Option<i32>,
+    pub ro: Option<u8>,
 
     // ----------------------------
     /// sender node id
@@ -36,7 +36,7 @@ impl FindNode {
         t: TransactionId,
         v: Option<Bytes>,
         ip: Option<SocketAddr>,
-        ro: Option<i32>,
+        ro: Option<u8>,
     ) -> Self {
         Self {
             id,
